refactor(frontend): split cover-image-upload into subcomponents (#113 Phase 1b)
Extract CircularProgress component into separate file.
Split 611-line component into focused modules:
- circular-progress.tsx: SVG progress indicator
- cover-image-upload.tsx: main upload component
- index.ts: barrel export

Import path unchanged via index.ts re-export.
