+++ frontend/src/components/ui/cover-image-upload/index.ts
@@ -1,0 +1,2 @@
+export { CoverImageUpload } from "./cover-image-upload";
+export { CircularProgress } from "./circular-progress";
